WIP: Adding homepage styling and population from API
