+++ src/models/AdExt.ts
@@ -1,21 +1,20 @@
-export type AdExt {
+export type AdExt = {
   id: number;
   headline: string;
   description: {
     text_formatted: string;
-  }
+  };
 
   employer: {
     name: string;
   };
 
-  workplace_address:{
+  workplace_address: {
     region: string;
     municipality: string;
-  }
+  };
 
   publication_date: string;
-
 
   application_contacts?: {
     description?: string;
@@ -23,4 +22,4 @@ export type AdExt {
     telephone?: string;
   };
   application_deadline: string;
-}
+};
