+++ src/pages/Layout.tsx
@@ -1,8 +1,8 @@
-import { Outlet } from 'react-router';
-import { Header } from '../components/Header';
-import { Footer } from '../components/Footer';
-import { DigiLayoutBlock } from '@digi/arbetsformedlingen-react';
-import { LayoutBlockVariation } from '@digi/arbetsformedlingen';
+import { Outlet } from "react-router";
+import { Header } from "../components/Header";
+import { Footer } from "../components/Footer";
+import { DigiLayoutBlock } from "@digi/arbetsformedlingen-react";
+import { LayoutBlockVariation } from "@digi/arbetsformedlingen";
 
 export const Layout = () => {
   return (
